+++ src/configs/memory-serializer.ts
@@ -1,5 +1,5 @@
-import { CodeplugReadResponse } from "../modules/radio-types/base";
-import { Codeplug } from "../proto/gen/cps/model/v1/codeplug_pb";
+import { CodeplugReadResponse } from "@modules/radio-types/base";
+import { Codeplug } from "@proto/cps/model/v1/codeplug_pb";
 
 export interface MemorySerializer {
   deserialize(buffer: CodeplugReadResponse): Codeplug;
