+++ src/configs/radios.tsx
@@ -1,4 +1,4 @@
-import { SerialRadio } from "../modules/radio-types/transports/serial/serial";
+import { SerialRadio } from "@modules/radio-types/transports/serial/serial";
 import { RadioDefinition } from "./radio-config";
 import { QuanshengUVK5Definition } from "./radios/quansheng/uvk5/radio";
 import { RaddiodityGD88Definition } from "./radios/raddiodity-gd88";
